+++ backend/src/main/java/com/senac/gestaocurso/strategy/funcionarioValidacaoImp/ImplementaListaDadoBancario.java
@@ -1,4 +1,4 @@
-package com.senac.gestaocurso.strategy.impl;
+package com.senac.gestaocurso.strategy.funcionarioValidacaoImp;
 
 import com.senac.gestaocurso.models.DadosBancarios;
 import com.senac.gestaocurso.models.Funcionario;
