+++ backend/src/main/java/com/senac/gestaocurso/strategy/funcionarioValidacaoImp/ValidarPisPasep.java
@@ -1,4 +1,4 @@
-package com.senac.gestaocurso.strategy.impl;
+package com.senac.gestaocurso.strategy.funcionarioValidacaoImp;
 
 import com.senac.gestaocurso.enterprise.ValidationException;
 import com.senac.gestaocurso.models.Funcionario;
@@ -8,19 +8,19 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @Component
-public class ValidarReservista implements NovaValidacaoFuncionarioStrategy {
+public class ValidarPisPasep implements NovaValidacaoFuncionarioStrategy {
 
     @Autowired
     private FuncionarioRepository repository;
 
     @Override
     public void validar(Funcionario funcionario) {
-        if (reservistaValidado(funcionario.getReservista())){
-            throw new ValidationException("Reservista já cadastrada");
+        if (pisPasepValidado(funcionario.getPisPasep())){
+            throw new ValidationException("PIS/PASEP já cadastrado");
         }
     }
 
-    private boolean reservistaValidado(String reservista){
-        return repository.findByReservista(reservista) != null;
+    private boolean pisPasepValidado(String pispasep){
+        return repository.findByPisPasep(pispasep) != null;
     }
 }
