+++ backend/src/main/java/com/senac/gestaocurso/strategy/funcionarioValidacaoImp/ValidarTituloEleitor.java
@@ -1,4 +1,4 @@
-package com.senac.gestaocurso.strategy.impl;
+package com.senac.gestaocurso.strategy.funcionarioValidacaoImp;
 
 import com.senac.gestaocurso.enterprise.ValidationException;
 import com.senac.gestaocurso.models.Funcionario;
@@ -8,19 +8,19 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @Component
-public class ValidarRg implements NovaValidacaoFuncionarioStrategy {
+public class ValidarTituloEleitor implements NovaValidacaoFuncionarioStrategy {
 
     @Autowired
     private FuncionarioRepository repository;
 
     @Override
     public void validar(Funcionario funcionario) {
-        if (rgValidado(funcionario.getRg())){
-            throw new ValidationException("RG já cadastrado");
+        if (tituloValidado(funcionario.getTituloEleitor())){
+            throw new ValidationException("Título de Eleitor já cadastrado");
         }
     }
 
-    private boolean rgValidado(String rg){
-        return repository.findByRg(rg) != null;
+    private boolean tituloValidado(String titulo){
+        return repository.findByTituloEleitor(titulo) != null;
     }
 }
